+++ Craftsman/Commands/AddFeatureCommand.cs
@@ -11,7 +11,9 @@
     using Spectre.Console;
     using Craftsman.Validators;
     using System.IO;
+    using System.Linq;
     using Craftsman.Builders.Features;
+    using Enums;
 
     public static class AddFeatureCommand
     {
@@ -32,7 +34,7 @@
             WriteHelpText(Environment.NewLine);
         }
 
-        public static void Run(string solutionDirectory)
+        public static void Run(string solutionDirectory, IFileSystem fileSystem)
         {
             try
             {
@@ -44,8 +46,17 @@
                 var contextName = Utilities.GetDbContext(srcDirectory, projectBaseName);
 
                 var feature = RunPrompt();
-                EmptyFeatureBuilder.CreateCommand(srcDirectory, contextName, projectBaseName, feature);
-
+                // EmptyFeatureBuilder.CreateCommand(srcDirectory, contextName, projectBaseName, feature);
+                EntityScaffolding.AddFeatureToProject(
+                    srcDirectory,
+                    testDirectory,
+                    projectBaseName,
+                    contextName,
+                    true,
+                    feature.Policies,
+                    feature,
+                    new Entity() { Name = feature.EntityName, Plural = feature.EntityPlural});
+                
                 WriteHelpHeader($"{Environment.NewLine}Your feature has been successfully added. Keep up the good work! {Emoji.Known.Sparkles}");
                 AnsiConsole.WriteLine();
             }
@@ -83,6 +94,34 @@
             AnsiConsole.WriteLine();
             AnsiConsole.Render(new Rule("[yellow]Add a New Feature[/]").RuleStyle("grey").Centered());
 
+            var featureType = AskFeatureType();
+            
+            if (featureType != FeatureType.AdHoc.Name)
+            {
+                var entityName = AskEntityName();
+                var entityPlural = AskEntityPlural(entityName);
+                var useSwagger = AskUseSwaggerComments();
+                var policies = AskPolicies();
+
+                AnsiConsole.WriteLine();
+                AnsiConsole.Render(new Table().AddColumns("[grey]Property[/]", "[grey]Value[/]")
+                    .RoundedBorder()
+                    .BorderColor(Color.Grey)
+                    .AddRow("[grey]Entity Name[/]", entityName)
+                    .AddRow("[grey]Entity Plural[/]", entityPlural)
+                    .AddRow("[grey]Use Swagger[/]", useSwagger.ToString())
+                    .AddRow("[grey]Policies[/]", policies.Count > 0 ? $"{policies.Count} policies" : "-")
+                );
+                
+                return new Feature()
+                {
+                    Type = featureType,
+                    EntityName = entityName,
+                    EntityPlural = entityPlural,
+                    Policies = policies
+                };
+            }
+            
             var feature = AskFeature();
             var command = AskCommand(feature);
             var responseType = AskResponseType();
@@ -97,9 +136,8 @@
                 .AddRow("[grey]Command Name[/]", command)
                 .AddRow("[grey]Response Type[/]", responseType)
                 .AddRow("[grey]Is Producer[/]", producer.ToString())
-                .AddRow("[grey]Entity Plural[/]", entityPluralForDir)
             );
-
+            
             return new Feature()
             {
                 Type = "AdHoc",
@@ -163,20 +201,102 @@
             return command;
         }
 
-        private static string AskSport(IAnsiConsole console)
+        private static string AskFeatureType()
         {
-            console.WriteLine();
-            console.Write(new Rule("[yellow]Choices[/]").RuleStyle("grey").LeftAligned());
-
-            return console.Prompt(
-                new TextPrompt<string>("What's your [green]favorite sport[/]?")
-                    .InvalidChoiceMessage("[red]That's not a sport![/]")
-                    .DefaultValue("Sport?")
-                    .AddChoice("Soccer")
-                    .AddChoice("Hockey")
-                    .AddChoice("Basketball"));
+            return AnsiConsole.Prompt(
+                new SelectionPrompt<string>()
+                    .Title("What [green]type of feature[/] do you want to add?")
+                    .PageSize(50)
+                    .AddChoices("AdHoc", "AddRecord", "DeleteRecord", "GetList", "GetRecord", "UpdateRecord", "PatchRecord"));
         }
 
+        private static string AskEntityName()
+        {
+            return AnsiConsole.Prompt(
+                new TextPrompt<string>("What's the [green]name of the entity[/] that will use this feature?")
+            );
+        }
+
+        private static string AskEntityPlural(string entityName)
+        {
+            return AnsiConsole.Prompt(
+                new TextPrompt<string>($"What's the [green]plural name[/] of the entity that will use this feature (Default: [green]{entityName}s[/])?")
+                    .DefaultValue($"{entityName}s")
+                    .HideDefaultValue()
+            );
+        }
+
+        private static bool AskUseSwaggerComments()
+        {
+            var command = AnsiConsole.Prompt(
+                new TextPrompt<string>("Do you want to include swagger comments? (Default: [green]y[/])?")
+                    .InvalidChoiceMessage("[red]Please respond 'y' or 'n'[/]")
+                    .DefaultValue("y")
+                    .HideDefaultValue()
+                    .AddChoice("y")
+                    .AddChoice("n"));
+
+            return command == "y";
+        }
+
+        private static List<Policy> AskPolicies()
+        {
+            var command = AnsiConsole.Prompt(
+                new TextPrompt<string>("Do you want to add a policy to this feature? (Default: [green]n[/])?")
+                    .InvalidChoiceMessage("[red]Please respond 'y' or 'n'[/]")
+                    .DefaultValue("n")
+                    .HideDefaultValue()
+                    .AddChoice("y")
+                    .AddChoice("n"));
+
+            var wantsToAddPolicies = command == "y";
+
+            if (wantsToAddPolicies)
+            {
+                var policies = new List<Policy>();
+                var firstPolicy = AskPolicy();
+                policies.Add(firstPolicy);
+
+                var additional = "y";
+                while (additional == "y")
+                {
+                    additional = AnsiConsole.Prompt(
+                        new TextPrompt<string>("Do you want to add any more policies? (Default: [green]n[/])?")
+                            .InvalidChoiceMessage("[red]Please respond 'y' or 'n'[/]")
+                            .DefaultValue("n")
+                            .HideDefaultValue()
+                            .AddChoice("y")
+                            .AddChoice("n"));
+
+                    var wantsAdditional = additional == "y";
+                    if (!wantsAdditional) break;
+                    
+                    var newPolicy = AskPolicy();
+                    policies.Add(newPolicy);
+                }
+
+                return policies;
+            }
+
+            return new List<Policy>();
+        }
+        
+        
+        private static Policy AskPolicy()
+        {
+            // ask policy type, then
+            var name = AskPolicyName();
+            return new Policy()
+            {
+                Name = name
+            };
+        }
+
+        private static string AskPolicyName()
+        {
+            return AnsiConsole.Ask<string>("What's the name of the policy you want to apply?");
+        }
+        
         private static int AskAge(IAnsiConsole console)
         {
             console.WriteLine();