+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 * Added BaseEntity that all entities will inherit from.
   * TODO: docs
+* Added built in features to the `add:feature` command
 
 ### Updated